reorganize temporary sched form into 1 step
